+++ src/features/rest-enhancements/index.js
@@ -8,7 +8,6 @@ import {logger} from '../../utils/logger';
 
 import { registerRestEnhancementsSettings } from './settings';
 import { handleShortRest } from './short-rest';
-import { handleCombatRecovery } from './combat-recovery';
 import { registerRestEnhancementsTests } from './quench';
 
 export function registerRestEnhancementsFeature() {
@@ -24,12 +23,6 @@ export function registerRestEnhancementsFeature() {
     Hooks.on('dnd5e.shortRest', async (actor, data) => {
         if (game.settings.get(module_id, 'rest-enhancements')) {
             await handleShortRest(actor, data);
-        }
-    });
-
-    Hooks.on('dnd5e.combatRecovery', async (combatant, recoveries) => {
-        if (game.settings.get(module_id, 'rest-enhancements')) {
-            await handleCombatRecovery(combatant, recoveries);
         }
     });
 }